chatbot_api/pyproject.toml
[project]
name = "chatbot_api"
version = "0.1"
dependencies = [
    "asyncio==3.4.3",
    "fastapi==0.109.0",
    "langchain==0.1.0",
    "langchainhub==0.1.14",
    "numpy==1.26.2",
    "opentelemetry-api==1.22.0",
    "pydantic==2.5.1",
    "python-dotenv==1.0.1",
    "uvicorn==0.25.0"
]

[project.optional-dependencies]
dev = ["black", "flake8"]

[tool.setuptools.packages.find]
where      = ["."]
include    = ["app"]
namespaces = true